requirements.txt
# Core Web App
fastapi>=0.111.0
uvicorn[standard]==0.29.0

# Database
SQLAlchemy==2.0.30
mysql==0.0.3
mysql-connector-python==9.4.0

# Pydantic
pydantic>=2.7.4
pydantic-settings>=2.4.0,<3.0.0
pydantic[email]

# JWT Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
PyJWT==2.10.1

# CORS & Form support
python-multipart==0.0.9

# File parsing
pdfplumber==0.11.0
PyMuPDF==1.23.17
pypdf==5.9.0

# LLM / OpenAI
openai==1.30.1
sentence_transformers==5.0.0
huggingface-hub==0.34.3
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9

# Vector search (FAISS)
faiss-cpu==1.11.0.post1
numpy==2.3.2

# Dev / Testing
pytest==8.2.2
httpx==0.27.0
Jinja2==3.1.6
